+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v4.63.0 Fixes
 
+- `[Accordion]` Added expand animation back. ([#6268](https://github.com/infor-design/enterprise/issues/6268))
 - `[Badges]` Fixed a bug where in badges is not properly aligned in Contrast Mode. ([#6273](https://github.com/infor-design/enterprise/issues/6273))
 - `[Button]` Fixed a bug where notification badges are not destroyed when updating the button settings. ([NG#1241](https://github.com/infor-design/enterprise-ng/issues/1241))
 - `[Calendar]` Allowed product devs to add custom css class to event labels in Calendar Component. ([#6304](https://github.com/infor-design/enterprise/issues/6304))